fix(subscriptions): The subscription resolvers were firing too often
The underlying subscription in a remote schema was not terminated properly, causing memory leaks. In
addition, multiple subscriptions caused duplicate payloads. Both is fixed by removing the PubSub
dependency and directly transforming the Observer returned by the ApolloLink to an async iterable.
This allows the async iterable to propagate to the underlying observable to cancel the suscription
once the subscription is not needed anymore.