Prevent visitSchema from replacing the original GraphQLSchema object.
Implementing the visitSchema method can be a good idea if you want to
modify the existing GraphQLSchema object, but replacing the entire schema
is pointless, because you're basically disregarding the result of
makeExecutableSchema.

Also, SchemaDirectiveVisitor.visitSchemaDirectives relies on the schema
object remaining the same throughout the traversal, and it would be tricky
to update the schema (as seen by visitSchemaDirectives) mid-traversal.