Simplify buildSchemaFromTypeDefinitions (no behavioral changes).
No longer using the concatenateTypeDefs function, because it needlessly
reprints DocumentNode objects just so that we can parse them again. I
would have removed the concatenateTypeDefs function completely, but it's
exported from the schemaGenerator.ts module, and tested, so that would be
a breaking change.

This code is also now able to de-duplicate function references regardless
of whether they appear as sibling elements in the same list.