feat(stitching): allow delegateToSchema, mergeSchemas and transformSchema to take remote schema configurations as parameters
This removes the need for makeRemoteExecutableSchema, removing an unnecessary layer of schema delegation.

This change introduces a RemoteGraphQLSchema type which consists of a GraphQLSchema object annotated with link, fetcher, or dispatcher properties that can be used by delegateToSchema to access the remote schema. The dispatcher function takes the graphql context eventually passed to delegateToSchema as an argument and returns a link or fetcher function.