fix(transforms): to properly allow chaining.
BREAKING CHANGE:

Result transforms should be reversed, so that multiple request and result transforms can be properly executed.

Request and result transforms must be inverted when transforming a schema, as these transforms work to move from the final schema back to the original schema.

These CHANGES have the potential to be BREAKING.