fix(stitching): change merged result format
Merged results that are null may carry errors from deeper within the query tree. Previous merged result format carries this metadata within a property on the result, but null has no properties.

New result format differs only in that a null result is transformed to an object with a special property signifying that the result was null, so that metadata can be added in the same way.

When merging within defaultMergedResolver, all null checks must check for this property as well. Because a result may be changed during annotation, the function is essentially no longer annotating only/i.e. modifying the object in place, and so it has been renamed to reflect that, with forEach changed to map when processing a list.

Fixes #26.