refactor(hoisting)
No need to prealias when wrapping if using deep merge instead of assign
to merge types from multiple schemas, as the query tree should always be
unique.

If not prealiasing when wrapping, wrapping does not require a delimeter,
and so hoisting does not require a prefix either as all delimeters are
used only when hoisting.