Inline `PrintSchemaOptions` type, whose parent module has been moved.
This option will likely be deprecated in `graphql@16`.  For now, we'll
inline this type into this module to continue emitting it into the
declaration file for `makeRemoteExecutableSchema`.

This is necessary since the TypeScript compiler can no longer resolve its
previous location as `graphql` has moved the location of the `schemaPrinter`
module to `printSchema` in https://github.com/graphql/graphql-js/pull/2426.

cc @IvanGoncharov