Remove defunct GitHunt-Api reference.
Since the repository at https://github.com/apollostack/GitHunt-API seems to have either been taken down or made private, and there does not appear to be a similar repository in the apollographql namespace, there probably shouldn't be examples pointing at it.

There seems to be many other references to the GitHunt-API repository in the official Apollo documentation as well as other 3rd party examples. If possible, could the repository be brought back? As a new user of GraphQL, I, and probably a lot of other new users, would appreciate seeing the example that so many tutorials refer to.