Fix stitching from and to interfaces (#1443)
* ExpandAbstractTypes to check transformed subschema

An abstract type may be present in the target schema, but renamed. ExpandAbstractTypes expands the abstract types not present in the target schema, but it should check the transformed target schema, to not be misled by renaming.

This may require manually passing the transformed schema in some cases. The default can stay to assume no renaming.

Within stitched schemas, the correct transformed schema can be saved to and then read from info.mergeInfo.transformedSchema.

* move ICreateRequest to delegate package with ICreateRequestFromInfo

* Add WrapConcreteTypes transform

= fixes #751

* allow redelegation of nested root query fields

in some circumstances

nested root fields may not always successfully stitch to subschemas, for example when the root field is a new  abstract field stitching to a concrete field or the reverse

* provide transformedSchema argument in more instances

= changes CreateResolverFn signature to take an options argument of type ICreateResolverOptions