remove need for fieldToFieldConfig function
BREAKING CHANGE: transformer functions will be passed field config objects instead of field objects.

fieldToFieldConfig was necessary because TransformRootField et al functions expected users to pass a function that receives a field and returns a field config, modified as desired. This can be made easier for end users by simply passing the transformer function a field config instead of a field, obviating the need for fieldToFieldConfig.