don't mutate previous schemas with HoistField and WrapFields transforms
fields helper functions now take a schema and return a new schema, rather than mutating the original schema's typemap.

part of broader push throughout the code base to avoid mutating schemas

Note: specifically within the context of transforms, we are currently protected against any side effects when mutating the passed in schema in any case. cloneSchema is used during the transformation loop -- the original schema cannot be mutated by a transformation, because a clone is passed rather than the original. Nevertheless, these changes are part of the overall effort to never rely on the ability to mutate an existing graphql schema.