Use field configs (#1451)
* remove spurious uses of fieldToFieldConfig

If type.toConfig() is later called returning config, fieldToFieldConfig is called every field to populate config.fields, so no need to call it individually from the field map, might as well just move the toConfig call earlier.

* remove need for fieldToFieldConfig function

BREAKING CHANGE: transformer functions will be passed field config objects instead of field objects.

fieldToFieldConfig was necessary because TransformRootField et al functions expected users to pass a function that receives a field and returns a field config, modified as desired. This can be made easier for end users by simply passing the transformer function a field config instead of a field, obviating the need for fieldToFieldConfig.

* replace mapSchemaByField with more MapperKind options within mapSchema

* hoist schema transformation out of type/field loop

this should be performed only once!

* don't mutate previous schemas with HoistField and WrapFields transforms

fields helper functions now take a schema and return a new schema, rather than mutating the original schema's typemap.

part of broader push throughout the code base to avoid mutating schemas

Note: specifically within the context of transforms, we are currently protected against any side effects when mutating the passed in schema in any case. cloneSchema is used during the transformation loop -- the original schema cannot be mutated by a transformation, because a clone is passed rather than the original. Nevertheless, these changes are part of the overall effort to never rely on the ability to mutate an existing graphql schema.

* remove use of createNamedStub

and add comment to explain how this works

* bring back TypeMap