add schema transforms
schema transform functions take a schema as an argument and return a schema, possibly modifying it

a schema transform function can be templated, i.e. one can create a function that takes certain arguments, such as the names of directives that annotate fields of interest, and return a schema transform function that modifies the schema based on the specified annotating directives

this allows one to reuse a schema transform function across projects, allowing for the customization of the schema transform function on a per use basis

schema transform functionc can be called directly on a schema to modify it before use -- makeExecutableSchema can now also be passed an array of schema transform functions, with makeExecutableSchema responsible for performing all schema transformations prior to returning.

Eventually, all schema visitor examples can be rewritten using the schema transform functional API

addresses #1234