avoid infinite loop (#1505)
fixSchemaAst fixes the schema ast by adding a getter for the property which lazily calls buildSchema(printSchemaWithDirectives(schema) to rebuild the schema.

printSchemaWithDirectives, however, now checks to see if schema.astNode is defined, and so will end up calling itself. This can be avoided by making sure that the astNode is not defined as a getter.

fixSchemaAst could alternatively be changed to return a new schema ratehr than modifying the original schema.