allow MapFields to transform all field nodes
MapFields (really should be MapFieldNodes, this should be changed in next major release) currently provides a quick way to modify field nodes for object types by specification of a map of type/field names as opposed to a function.

this backwards-compatible change allows MapFields to work for interfaces and root types.