allow mutation type wrapping
the blocker for this was that the default createProxyingResolverFn was changed to no longer hardcode the operation, but to rather infer it from the root type being stitched from.

This was in parallel to a similar change with the target fieldName which was helpful when making the RenameRootField transformer just a special case of the RenameObjectField transformer, but messes up the wrapping of mutations, because when the root fields in the wrapping subschema are the ones that actually do the proxying, and wrapping them with another type causes them to default to the query operation.

So, the default createProxyingResolverFn now hardcodes the operation, and all is well.