fix(createRequest): selectionSet overrides fieldNodes
selectionSet option is used with bare execution, where fieldNodes is not present, without issue.

with type merging, info.fieldNodes and fieldNodes may be specified, but selectionSet should override, as there may be situation when both appear but selectionSet desired, unlikely for them both to appear but fieldNodes to be desired, as fieldNodes is auto generated, and selectionSet is more likely to be created manually.