enhance(typeMerging): prune fields that cause key fields to be added
key fields for a given subschema by definition must be resolved outside that subschema, and so for any given key field, there is no need to add the other key fields to the selectionSet or to consider the key field as proxiable via that subschema.

The key field is the "key" to the remaining fields, but -- for the purposes of type merging -- will never be obtained from the given subschema.

This becomes more complex with non-leaf key fields, and so this pruning is enabled only for leaf key fields.