works when batching is disabled, so this is a bug in batching not with fieldSelectionSet isolation
On the other hand, will ignore the price and weight given by its own query in favor of the Product service, so may give different results when queried directly and when queried via Gateway.

So isolation definitely needs to go behind a flag as opt-in. Name of flag might be `fieldSelectionSetsViaGateway` or something like that to describe the different behavior rather than the internal implementation. Anything sound catchier?