fix(HoistField): allow hoisting to root
requires passing additional parameters to each transformSchema call beyond the subschemaConfig: the transforms, and the transformedSchema.

Because transformSchema may need to create additional proxying resolvers, it needs all the parameters that would be passed to delegateToSchema.

It is NOT a catch-22 that transformSchema requires the transformed schema as a parameter, because it does NOT require an executable transformedSchema as a parameter, only the non-executable version for type comparisions performed ultimately by delegateToSchema within the proxying process.