BREAKING_CHANGE: remove custom wrappingResolver option from WrapFields
This does not make sense in the context of planned functionality to wrap subscription root fields in potentially multiple layers, as the wrapping resolvers will have to be different for the root field and the other layers.

Changing a root field resolver can be easily accomplished in a separate transform via TransformRootFIelds or TransformObjectFIelds.