[deploy_website] enhance(stitch): custom merge resolvers (#2171)
Upgrade MergedTypeConfig key function to intrinsic part of merge algorithm. Previously, merge resolvers had to manually call the key function, now, if a key function is specified, it is passed as additional argument to the merge resolver.

Export new function createMergedTypeResolver that takes the relevant subset of MergedTypeConfig and returns the default resolver.

The resolve field can therefore be left blank and will automatically be created by createMergedTypeResolver or it can be set by end users who can wrap the default resolver as they wish.

New type MergedTypeResolverOptions represents the subset of config that can be used with createMergedTypeResolver.

Fixes bugs in tests that used args with batching instead of argsFromKeys, uncovered by the refactoring required for these changes.

Adds docs.