Fix Error When Loading Schema URLs Synchronously
The `sync-fetch` library does not properly support aborting Fetch
requests, so when attempting to call `UrlLoader#loadSync`, you get an
error stating that the signal you passed in is not an instance of
`AbortSignal` (it actually is, but there's some kind of problem with how
sync-fetch is sending this over or how node-fetch is reading
it...honestly it's a bit of a rabbit hole).

This change ensures that the only time a `signal:` is sent to fetch is
when the fetch implementation can handle it. If `sync-fetch` is used to
implement the Fetch API, a signal should not be sent over since it is
not capable of aborting requests. Downstream tooling like graphql-config
and graphql-eslint will be able read URLs just as they do with schema
paths and other schema sources.

Fixes #3174