Allow empty selection set in TransformQuery#queryTransformer
806faae3c7189641de8a1bfb4bcfcfba95240e47 turned on strict mode, which
changed the meaning of the original type annotation for
QueryTransformer (turning SelectionSetNode into a non-nullable type),
and then the implementation was changed to match the _new_ type.

That means that the above commit introduced an API change where
queryTransformer would not get called when there's no selectionSet in
the original query. Let's change that back (if the maintainer agrees).