Create symbols with Symbol.for()
When multiple copies of `@graphql-tools/delegate` are available in a workspace (eg: because `stitch` and `wrap` depend on different versions), creating symbols with `Symbol()` can lead to bugs when one version of `delegate` sets an object property using its version of the symbol which the other `delegate` instance will not be able to access. Using `Symbol.for()` fixes this: as long as both versions of delegate use the same string description, they will refer to the same symbol.